requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
apache-airflow==3.1.8
apache-airflow-core==3.1.8
apache-airflow-providers-common-compat==1.14.2
apache-airflow-providers-common-io==1.7.2
apache-airflow-providers-common-sql==1.34.0
apache-airflow-providers-smtp==2.4.4
apache-airflow-providers-standard==1.12.2
apache-airflow-task-sdk==1.1.8
argcomplete==3.6.3
asgiref==3.10.0
attrs==26.1.0
babel==2.18.0
blinker==1.9.0
cachetools==6.2.2
cadwyn==5.4.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cron_descriptor==2.0.8
croniter==6.2.2
cryptography==46.0.6
Deprecated==1.3.1
dill==0.4.1
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.24
Flask==3.1.2
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenback==1.3.0
greenlet==3.3.2
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.1
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.1.0
lockfile==0.12.2
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
methodtools==0.4.7
more-itertools==11.0.1
msgspec==0.20.0
mysql-connector-python==9.6.0
natsort==8.4.0
numpy==2.4.2
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.1
pathspec==1.0.4
pendulum==3.2.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.2.2
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.20.0
pygtrie==2.5.0
PyJWT==2.12.1
PyMuPDF==1.26.6
pyparsing==3.2.5
PyPDF2==3.0.1
pyspark==3.5.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.24
python-slugify==8.0.4
PyYAML==6.0.3
PyYAML-ft==8.0.0
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rich-argparse==1.7.2
rich-toolkit==0.19.7
rpds-py==0.30.0
rsa==4.9.1
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.3
starlette==0.48.0
structlog==25.5.0
svcs==25.1.0
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
tqdm==4.67.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==2.0.0
universal_pathlib==0.2.6
uritemplate==4.2.0
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==2.1.2
zipp==3.23.0
